day6 v2

+++ srcs/apply_length.c
@@ -16,9 +16,9 @@ int		apply_length_d(t_format *format, va_list ap)
 {
 	if (!ft_strcmp(format->length_flag, "h"))
 		return (convert_int2string(format, \
-			(int)((short)(va_arg(ap, int))), 10));
+			(short)(va_arg(ap, int)), 10));
 	else if (!ft_strcmp(format->length_flag, "hh"))
-		return (convert_int2string(format, (int)((char)(va_arg(ap, int))), 10));
+		return (convert_int2string(format, (char)(va_arg(ap, int)), 10));
 	else if (!ft_strcmp(format->length_flag, "l"))
 		return (convert_int2string(format, va_arg(ap, long), 10));
 	else if (!ft_strcmp(format->length_flag, "ll"))
@@ -30,9 +30,9 @@ int		apply_length_u(t_format *format, va_list ap)
 {
 	if (!ft_strcmp(format->length_flag, "h"))
 		return (convert_int2string(format, \
-			(int)((short)(va_arg(ap, int))), 10));
+			(short)(va_arg(ap, int)), 10));
 	else if (!ft_strcmp(format->length_flag, "hh"))
-		return (convert_int2string(format, (int)((char)(va_arg(ap, int))), 10));
+		return (convert_int2string(format, (char)(va_arg(ap, int)), 10));
 	else if (!ft_strcmp(format->length_flag, "l"))
 		return (convert_int2string(format, va_arg(ap, unsigned long), 10));
 	else if (!ft_strcmp(format->length_flag, "ll"))
@@ -44,9 +44,9 @@ int		apply_length_x(t_format *format, va_list ap)
 {
 	if (!ft_strcmp(format->length_flag, "h"))
 		return (convert_intXO2string(format, \
-			(int)((short)(va_arg(ap, unsigned int))), 16));
+			(short)(va_arg(ap, unsigned int)), 16));
 	else if (!ft_strcmp(format->length_flag, "hh"))
-		return (convert_intXO2string(format, (int)((char)(va_arg(ap, unsigned int))), 16));
+		return (convert_intXO2string(format, (char)(va_arg(ap, unsigned int)), 16));
 	else if (!ft_strcmp(format->length_flag, "l"))
 		return (convert_intXO2string(format, va_arg(ap, unsigned long), 16));
 	else if (!ft_strcmp(format->length_flag, "ll"))
@@ -57,9 +57,9 @@ int		apply_length_x(t_format *format, va_list ap)
 int		apply_length_o(t_format *format, va_list ap)
 {
 	if (!ft_strcmp(format->length_flag, "h"))
-		return (convert_intXO2string(format, (int)((short)(va_arg(ap, unsigned int))), 8));
+		return (convert_intXO2string(format, (short)(va_arg(ap, unsigned int)), 8));
 	else if (!ft_strcmp(format->length_flag, "hh"))
-		return (convert_intXO2string(format, (int)((char)(va_arg(ap, unsigned int))), 8));
+		return (convert_intXO2string(format, (char)(va_arg(ap, unsigned int)), 8));
 	else if (!ft_strcmp(format->length_flag, "l"))
 		return (convert_intXO2string(format, va_arg(ap, unsigned long), 8));
 	else if (!ft_strcmp(format->length_flag, "ll"))
